+++ packages/core/utils/bar/prepareDrink.ts
@@ -1,3 +1,5 @@
+import chalk from 'chalk';
+
 const getName = (name: string) => {
   if (!name) {
     return '.';
@@ -12,13 +14,12 @@ export const prepareDrink = (name: string) => {
     setTimeout(() => {
       console.log('🍸 Adding gin and lime juice...');
       setTimeout(() => {
-        console.log('🍸 Topping with', '\x1b[34mTonik\x1b[0m...');
+        console.log('🍸 Topping with', chalk.blue('Tonik') + '...');
         setTimeout(() => {
           console.log('🍸 Garnishing with lime wedge...');
           setTimeout(() => {
-            console.log(`\x1b[32m%s\x1b[0m`, `🍸 Your Stapled ${getName(name)} is ready!`, `\x1b[0m`);
-            // I'm too lazy to mess with modules building to allow coloring library to be installed lol
-            console.log(`🍸 You can now run:`, `\x1b[36mcd ${name} && pnpm dev\x1b[0m`);
+            console.log(chalk.green(`🍸 Your Stapled ${getName(name)} is ready!`));
+            console.log(`🍸 You can now run:`, chalk.cyan(`cd ${name} && pnpm dev`));
           }, 1000);
         }, 1000);
       }, 1000);