+++ packages/core/utils/github/ghInstaller.ts
@@ -1,5 +1,6 @@
 import { execSync } from 'child_process';
 import * as os from 'os';
+import { getLogColor } from '../shared/getLogColor';
 
 export const isGitHubCLIInstalled = (): boolean => {
   try {
@@ -27,8 +28,10 @@ export const installGitHubCLI = (): boolean => {
         installCommand =
           'sudo dnf config-manager --add-repo https://cli.github.com/packages/rpm/gh-cli.repo && sudo dnf install gh';
       } else {
-        console.log('🖇️  Automatic installation is not supported for your Linux distribution.');
-        console.log('🖇️  Please visit https://github.com/cli/cli#installation for installation instructions.');
+        getLogColor('github', [
+          'Automatic installation is not supported for your Linux distribution.',
+          '\n Please visit https://github.com/cli/cli#installation for installation instructions.',
+        ]);
         return false;
       }
       break;
@@ -36,19 +39,21 @@ export const installGitHubCLI = (): boolean => {
       installCommand = 'winget install --id GitHub.cli';
       break;
     default:
-      console.log('🖇️  Automatic installation is not supported for your operating system.');
-      console.log('🖇️  Please visit https://github.com/cli/cli#installation for installation instructions.');
+      getLogColor('github', [
+        'Automatic installation is not supported for your operating system.',
+        '\nPlease visit https://github.com/cli/cli#installation for installation instructions.',
+      ]);
       return false;
   }
 
-  console.log('🖇️  Installing GitHub CLI...');
+  getLogColor('github', 'Installing GitHub CLI...');
   try {
     execSync(installCommand, { stdio: 'inherit' });
-    console.log('🖇️  GitHub CLI installed successfully.');
+    getLogColor('github', 'GitHub CLI installed successfully.');
     return true;
   } catch (error) {
-    console.error('🖇️  Failed to install GitHub CLI.');
-    console.log('🖇️  Please install it manually from: https://github.com/cli/cli#installation');
+    console.error('Failed to install GitHub CLI.');
+    getLogColor('github', 'Please install it manually from: https://github.com/cli/cli#installation');
     return false;
   }
 };