+++ packages/core/utils/supabase/install.ts
@@ -4,57 +4,58 @@ import path from 'path';
 import { supabaseFiles } from '../../templates/supabase/installConfig';
 import { templateGenerator } from '../generator/generator';
 import { getTemplateDirectory } from '../shared/getTemplateDirectory';
+import { getLogColor } from '../shared/getLogColor';
 
 const supabaseLogin = () => {
-  console.log('🖇️  Logging into Supabase...');
+  getLogColor('supabase', 'Logging in...');
 
   try {
     execSync('npx supabase projects list', { stdio: 'ignore' });
-    console.log('🖇️  Already logged into Supabase. Skipping login.');
+    getLogColor('supabase', 'Already logged in.');
     return;
   } catch (error) {
     try {
       execSync('npx supabase login', { stdio: 'inherit' });
     } catch {
-      console.error('\n🖇️  Failed to log in to Supabase.');
-      console.log('\n🖇️  Please log in manually with "supabase login" and re-run "create-stapler-app".');
+      console.error('Failed to log in to Supabase.');
+      getLogColor('supabase', '\nPlease log in manually with "supabase login" and re-run "create-stapler-app".');
       process.exit(1);
     }
   }
 };
 
 const initializeSupabaseProject = (): void => {
-  console.log('🖇️  Initialize Supabase project...');
+  getLogColor('supabase', 'Initialize project...');
   try {
     execSync(`npx supabase init`, { stdio: ['pipe'], encoding: 'utf-8' });
   } catch (error: any) {
     const errorMessage = error.stderr;
 
     if (errorMessage.includes('file exists')) {
-      console.log('\n🖇️  Supabase configuration file already exists. Skipping re-initialization.');
+      getLogColor('supabase', 'Configuration file already exists.');
       return;
     } else {
-      console.error('\n🖇️  Failed to initialize Supabase project with "supabase init".');
-      console.log(
-        '\n🖇️  Please review the error message below, follow the initialization instructions, and try running "create-stapler-app" again.',
-      );
-      process.exit(1);
+      console.error('Failed to initialize Supabase project with "supabase init".');
+      getLogColor(
+        'supabase',
+        '\nPlease review the error message below, follow the initialization instructions, and try running "create-stapler-app" again.',
+      ),
+        process.exit(1);
     }
   }
 };
 
 export const installSupabase = async (destinationDirectory: string) => {
-  console.log('🖇️  Installing supabase-js...');
+  getLogColor('supabase', 'Installing supabase-js...');
   try {
     supabaseLogin();
     initializeSupabaseProject();
   } catch (error) {
-    console.error('\n🖇️  Failed to init Supabase project.');
-    console.error(`🖇️  Error: ${error}`);
+    console.error('Failed to init project.', `\nError: ${error}`);
     process.exit(1);
   }
 
-  console.log('🖇️  Adding Supabase Files...');
+  getLogColor('supabase', 'Adding Files...');
 
   const templateDirectory = getTemplateDirectory(`/templates/supabase/files/`);
 
@@ -72,9 +73,9 @@ export const installSupabase = async (destinationDirectory: string) => {
 
   process.chdir('supabase');
 
-  console.log('🖇️  Installing Supabase dependencies...');
+  getLogColor('supabase', 'Installing dependencies...');
 
-  execSync('pnpm install', { stdio: 'inherit' });
+  execSync('pnpm i --reporter silent', { stdio: 'inherit' });
 
   process.chdir('..');
 };