+++ packages/core/utils/turbo/create.ts
@@ -1,7 +1,8 @@
 import { execSync } from 'child_process';
+import { getLogColor } from '../shared/getLogColor';
 
 export const createTurboRepo = async (name: string) => {
-  console.log(`🖇️  Creating Turbo project "${name}"...`);
+  getLogColor('turborepo', 'Initializing...');
   execSync(`npx create-turbo@latest ${name} -m pnpm`, {
     stdio: 'inherit',
   });